use the correct dtype in `ipex.optimize()` and use `set_seed` (#100)
* use set_seed

* change dtype

* change dtype

* remove enable

* add batch size note

* optimize README

* remove device

* use bf16 by default

* fix 4 to 6

* make ONEAPI ROOT custmizable

* no autocast_dtype

+++ testcase/automatic-speech-recognition/run_automatic-speech-recognition.py
@@ -81,7 +81,7 @@ if __name__ == "__main__":
             import intel_extension_for_pytorch as ipex
 
             logging.info("Use ipex optimize")
-            generator.model = ipex.optimize(generator.model, dtype=dtype, inplace=True)
+            generator.model = ipex.optimize(generator.model, dtype=torch_dtype, inplace=True)
 
         generate(
             generator, data["train"][0]["audio"]["array"], warm_up_steps, run_steps
